fix: row limits & row count labels are confusing
Currently, in explore & dashboard, we usually apply a row limit
on the query we issue against the database.

Following this, some visualization backend code does some
post-processing on data frames using pandas, typically pivoting
some things, which affects the result set's row count in
intricate capacities.

Now from a UX standpoint the user is exposed with:
- row limits in the control panels
- row count in various areas of the UI where visualiztion, preview,
  samples and raw results are shown

Currently we show the rowcount that's from row-processing. So maybe
a hard limit was applied at 1000 rows on the database, but we
pivot and it goes does to say 532, and we show the user that
we haven't hit the limit when we actually did.

Also note that the component that shows rowcount is supposed
to turn red if limit is hit, letting the user know they are looking at
truncated data.